+++ docs/history.rst
@@ -4,7 +4,10 @@ Changelog
 2.0.6 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix issue where indexes in settings aren't searched.
+  [instification]
+- Allow boost to be added to indexes in the settings control panel.
+  [instification]
 
 
 2.0.5 (2018-12-20)